Use portable unsigned int pointer types (USize/uintptr_t).

+++ lib/pony-aa.c
@@ -1,18 +1,18 @@
 
 #include <stdint.h>
 
-uint64_t pony_aa_obj_id(void *ptr) {
-  return (uint64_t)ptr;
+uintptr_t pony_aa_obj_id(void *ptr) {
+  return (uintptr_t)ptr;
 }
 
-void *pony_aa_obj_from_id(uint64_t id) {
+void *pony_aa_obj_from_id(uintptr_t id) {
   return (void *)id;
 }
 
 static void *s_ptr = 0;
 
-uint64_t pony_aa_get_main_id() {
-  return (uint64_t)s_ptr;
+uintptr_t pony_aa_get_main_id() {
+  return (uintptr_t)s_ptr;
 }
 
 void *pony_aa_get_main() {
